rest/connexion-template/pyproject.toml
[tool.poetry]
name = "connexion-template"
version = "0.0.0"
description = "Connexion REST API template"
authors = ["ChangZhuo Chen (陳昌倬) <czchen@czchen.org>"]
license = "MIT"

[tool.poetry.dependencies]
connexion = {version = "3.0.1", extras = ["flask", "swagger-ui", "uvicorn"]}
flask-migrate = "4.0.4"
flask-sqlalchemy = "3.0.5"
gunicorn = "21.2.0"
hydra-core = "1.3.2"
opentelemetry-instrumentation-flask = "0.40b0"
python = "^3.12"
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
testcontainers = {version = "3.7.1", extras = ["postgresql"]}
tox = "4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
